feat: add send modal in desktop

+++ src/ui/views/SendToken/index.tsx
@@ -80,7 +80,9 @@ import { TDisableCheckChainFn } from '@/ui/component/ChainSelector/components/Se
 import { useClearMiniGasStateEffect } from '@/ui/hooks/miniSignGasStore';
 
 const isTab = getUiType().isTab;
-const getContainer = isTab ? '.js-rabby-popup-container' : undefined;
+const isDesktop = getUiType().isDesktop;
+const getContainer =
+  isTab || isDesktop ? '.js-rabby-popup-container' : undefined;
 
 const abiCoder = (abiCoderInst as unknown) as AbiCoder;
 
@@ -306,7 +308,7 @@ const SendToken = () => {
   }, [currentAccount?.type]);
 
   useEffect(() => {
-    if (!toAddress) {
+    if (!toAddress && !getUiType().isDesktop) {
       const query = new URLSearchParams(search);
       query.delete('to');
       history.replace(
@@ -1542,19 +1544,19 @@ const SendToken = () => {
       <div
         className={clsx(
           'send-token',
-          isTab
+          isDesktop
             ? 'w-full h-full overflow-auto min-h-0 rounded-[16px] shadow-[0px_40px_80px_0px_rgba(43,57,143,0.40)'
             : ''
         )}
       >
         <PageHeader
           onBack={handleClickBack}
-          forceShowBack={!isTab}
-          canBack={!isTab}
-          isShowAccount
+          forceShowBack={!isDesktop}
+          canBack={true}
+          isShowAccount={!isDesktop}
           className="mb-[10px]"
           rightSlot={
-            isTab ? null : (
+            isDesktop ? null : (
               <div
                 className="text-r-neutral-title1 cursor-pointer absolute right-0 top-1/2 -translate-y-1/2"
                 onClick={() => {
@@ -1588,7 +1590,16 @@ const SendToken = () => {
                 <div
                   className="cursor-pointer text-r-neutral-title1"
                   onClick={() => {
-                    history.replace(`/send-poly${history.location.search}`);
+                    const query = new URLSearchParams(history.location.search);
+                    query.set('sendPageType', 'sendPoly');
+                    query.set('action', 'send');
+                    if (getUiType().isDesktop) {
+                      wallet.openInDesktop(
+                        `desktop/profile?${query.toString()}`
+                      );
+                    } else {
+                      history.replace(`/send-poly?${query.toString()}`);
+                    }
                   }}
                 >
                   <RcIconSwitchCC width={20} height={20} />
